fix(gui): move the disclosure hook out of the component file; tighten the new tests
CI's `static` lane failed on react-refresh/only-export-components: the file
exports components, `allowConstantExport` covers the class constant it
already exported, and a hook is neither. Run with `--max-warnings 0`, so the
warning is fatal. Move `useInterviewOptionDetailsDisclosure` and its
interface to their own module, which is what the rule asks for.

Test cleanup from review: the four disclosure tests shared one inline
fixture, and their option description read `Beta details` - the same string
as the trigger's accessible name (`${label} details`). The assertions could
not distinguish the revealed text from the button that reveals it if a query
ever started matching accessible names. Extract `renderDetailsCard` and
`tapWithFinger`, and give the description a distinct value.

Re-verified the tests still fail against the pre-fix source.
